+++ code/spikes/publish.001/src/Pkg.Content.Bundle/ContentBundle.TEST.mts
