+++ code/spikes/publish.001/src/Pkg.Content.Bundle/common.mts
