+++ code/spikes/publish.001/src/Pkg.Content.Bundle/types.mts
