+++ code/spikes/publish.001/src/Pkg.Content.Log/ContentLog.mts
