+++ code/spikes/publish.001/src/Pkg.Content.Log/index.mts
