+++ code/spikes/publish.001/src/types.mts
@@ -1,5 +1,5 @@
-export * from './Content.Bundle/types.mjs';
-export * from './Content.Log/types.mjs';
+export * from './Pkg.Content.Bundle/types.mjs';
+export * from './Pkg.Content.Log/types.mjs';
 export * from './Markdown.File/types.mjs';
 
 export * from './ui/state/types.mjs';
